+++ justinvest/portfolio.py
@@ -60,6 +60,9 @@ Each version is given a distinguishing version number. If the Library as you rec
 If the Library as you received it specifies that a proxy can decide whether future versions of the GNU Lesser General Public License shall apply, that proxy's public statement of acceptance of any version is permanent authorization for you to choose that version for the Library.
 '''
 import datetime
+import pandas as pd
+import requests
+import yfinance as yf
 from .rulesets import set_leverage
 
 def valid_date(date_text):
@@ -74,229 +77,300 @@ def valid_date(date_text):
 
 class Portfolio:
     '''
-    Portfolio class is the portfolio of the user. 
+    Portfolio class is a portfolio of the user. 
     User interacts with the class to construct and update portfolio.
+
+    ...
+
+    Attributes 
+    balance : int or float
+        a formatted string to print out what the animal says
+        
+    Methods
+    -------
+    deposit(deposit)
     '''
-    def __init__(self, balance):
-        self.balance = balance
+    def __init__(self, cash):
+        '''
+        User must start the instance with a balance
+        
+        ...
+
+        Parameters
+        ----------
+        balance : int or float
+        '''
+        assert isinstance(cash, (int, float)), "balance parameter must be Int or Float."
+        self.balance = cash
 
     def deposit(self, deposit):
-        self.balance += deposit
+        '''
+        Increase cash with a deposit
+
+        ...
+
+        Attributes
+        ----------
+        deposit : int or float
+        '''
+        assert isinstance(deposit, (int, float)), "deposit parameter must be Int or Float."
+        self.cash += deposit
     
     def withdraw(self, withdraw):
-        self.balance -= withdraw
-    
-    def get_leverage(self):
-        self.leverage = set_leverage(self.balance)
-        print(f"Current Leverage: {self.leverage}")
+        '''
+        Decrease cash with a deposit
 
-    def get_investable(self):
-        return self.balance * self.leverage
+        ...
 
-    def get_symbol(symbol):
-        url = "http://d.yimg.com/autoc.finance.yahoo.com/autoc?query={}&region=1&lang=en".format(symbol)
-
-        result = requests.get(url).json()
-
-        for x in result['ResultSet']['Result']:
-            if x['symbol'] == symbol:
-                return x['name']
-
-def getport(start, end, symbols):
-  '''
-  A function that collects all data from a list of assets from Yahoo! Finance
-  '''
-  assets = [get_symbol(x) for x in symbols]
-  valid_date(start)
-  valid_date(end)
-  assert isinstance(symbols, list), '代號必須是list of string'
-  for symbol in symbols:
-      assert isinstance(symbol, str), '代號必須是list of string'
-  assert isinstance(assets, list), '資產名稱必須是list of string'
-  for asset in assets:
-      assert isinstance(asset, str), '資產名稱必須是list of string'
-  assert len(symbols) == len(assets), 'symbols 和assets 必須是同數量'
-  # Define date_range
-  date_range = pd.date_range(
-    start = start,
-    end = end,  
-    freq = 'B', 
-    normalize = True
-    )
-  df_list = [yf.Ticker(symbol).history(start=start, end=end) for symbol in symbols]
-  for df, asset in zip(df_list, assets):
-    col_list = df.drop('Close', axis = 1).columns
-    df.drop(col_list, axis=1, inplace=True)
-    df.rename(columns={'Close': asset}, inplace=True)
-  df = pd.concat(df_list, axis=1, ignore_index=False)
-  df = df.fillna(method='ffill').reindex(index=date_range, method='ffill')
-  df.dropna(how='any', inplace=True)
-  return df
-
-def EfficientFrontier(table, annualized):
-    ### Populate portfolios
-# Calculate daily and annual returns of the stocks
-    assert isinstance(table, pd.DataFrame), "table 必須是DataFrame 格式"
-    assert isinstance(annualized, bool), "annualized 必須是True 或是False"
-    if annualized == False:
-      returns_daily = table.pct_change()
-      returns_total = (table.iloc[-1]-table.iloc[0])/table.iloc[0]
-
-      # Get daily and covariance of returns of the stock
-      cov_daily = returns_daily.cov()
-      cov_total = cov_daily * table.shape[0]
-      
-    elif annualized == True:
-      returns_daily = table.pct_change()
-      returns_total = returns_daily.mean() * 252
-
-      # Get daily and covariance of returns of the stock
-      cov_daily = returns_daily.cov()
-      cov_total = cov_daily * 252
-    else:
-      raise ValueError("annualized 必須是True 或是False")
-
-    # Define selected assets
-    selected = table.columns
-
-    # Empty lists to store returns, volatility and weights of imiginary portfolios
-    port_returns = []
-    port_volatility = []
-    sharpe_ratio = []
-    stock_weights = []
-
-    # Set the number of combinations for imaginary portfolios
-    num_assets = len(selected)
-    num_portfolios = 50000
-
-    # Set random seed for reproduction's sake
-    np.random.seed(101)
-
-    # Populate the empty lists with each portfolios returns,risk and weights
-    for single_portfolio in range(num_portfolios):
-        weights = np.random.random(num_assets)
-        weights /= np.sum(weights)
-        returns = np.dot(weights, returns_total)
-        volatility = np.sqrt(np.dot(weights.T, np.dot(cov_total, weights)))
-        sharpe = returns / volatility
-        sharpe_ratio.append(sharpe)
-        port_returns.append(returns)
-        port_volatility.append(volatility)
-        stock_weights.append(weights)
-
-    # A dictionary for Returns and Risk values of each portfolio
-    portfolio = {'回報': port_returns,
-                '波動': port_volatility,
-                '夏普比例': sharpe_ratio}
-
-    # Extend original dictionary to accomodate each ticker and weight in the portfolio
-    for counter,symbol in enumerate(selected):
-        portfolio[symbol+' 權重'] = [Weight[counter] for Weight in stock_weights]
-
-    # Make a nice dataframe of the extended dictionary
-    df = pd.DataFrame(portfolio)
-
-    # Get better labels for desired arrangement of columns
-    column_order = ['回報', '波動', '夏普比例'] + [stock+' 權重' for stock in selected]
-
-    # Reorder dataframe columns
-    df = df[column_order]
-    df.dropna(how="any", inplace=True, axis=0)
-    min_volatility = df['波動'].min()
-    max_sharpe = df['夏普比例'].max()
-
-    # use the min, max values to locate and create the two special portfolios
-    sharpe_portfolio = df.loc[df['夏普比例'] == max_sharpe]
-    min_variance_port = df.loc[df['波動'] == min_volatility]
-
-    mv_port = min_variance_port.copy()
-    ms_port = sharpe_portfolio.copy()
-    for col in min_variance_port.columns:
-        if col == '夏普比例':
-            pass
-        else:
-            mv_port[col] = mv_port[col].apply(lambda x: x*100)
-            
-    for col in sharpe_portfolio.columns:
-        if col == '夏普比例':
-            pass
-        else:
-            ms_port[col] = ms_port[col].apply(lambda x: x*100)
-    # plot frontier, max sharpe & min Volatility values with a scatterplot
-    fig = px.scatter(
-        df, 
-        x="波動", 
-        y="回報",  
-        size_max=60,
-        color="夏普比例",
-        color_continuous_scale = 'aggrnyl',
-        hover_data=df.columns
+        Attributes
+        ----------
+        deposit : int or float
+        '''
+        assert isinstance(withdraw, (int, float)), "withdraw parameter must be Int or Float."
         
-    )
-    fig.add_scatter(
-        x=sharpe_portfolio['波動'], 
-        y=sharpe_portfolio['回報'],
-        mode="markers",
-        marker={'size': 20,'color': "red",},
-        showlegend=False
-    )
-    fig.add_scatter(
-        x=min_variance_port['波動'], 
-        y=min_variance_port['回報'],
-        mode="markers",
-        marker={'size': 20,'color': "blue",},
-        showlegend=False
-    )
-    fig.update_traces(textposition='top center')
+        self.cash -= withdraw
+    
 
-    fig.update_layout(
-        height=600,
-        width=1000,
-        title_text='Portfolio'
-    )
-    pred_port = pd.concat([mv_port.T, ms_port.T], axis = 1)
-    pred_port.columns = ['最低波動', '最高夏普']
-    growth = (((table.iloc[-1]-table.iloc[0])/table.iloc[0])*100)
-    return fig, pred_port, growth
 
-# Codes modified from: https://medium.com/python-data/efficient-frontier-portfolio-optimization-with-python-part-2-2-2fe23413ad94
+    def search(self, *args):
+        '''
+        Queries result from Yahoo! Finance autocomplete API and returns a Pandas Dataframe.
+        Used to search for symbols.
 
-# 定義開始日期
-start = '2010-01-01'
+        ----
+        
+        '''
+        # A list of results
+        results = []
 
-# 定義結束日期
-end = '2019-12-31'
+        # For every query provided
+        for query in args:
 
-# 投資組合裡面的所有資產，必須是Yahoo! Finance 英文版上的代號，必須是在英文的引號("")裡面
-# 請上https://finance.yahoo.com/ 查詢代號
-symbols = ["GOOG",
-           "AAPL",
-           "FB",
-           "AMZN",
-           "TWD=X",
-           "^IRX",
-           "^FVX",
-           "^TNX",
-           "^TYX",
-           "WTI",
-           "USO",
-           "GLDM"
-           ]
+            # Query with the keyword
+            url = f"http://d.yimg.com/autoc.finance.yahoo.com/autoc?query={query}&region=1&lang=en"
 
-table = getport(start, end, symbols)
+            # Get the result in JSON format
+            result = requests.get(url).json()
 
-# annualized 是年化與否
-# True 會以年化平均計算，提供未來配置的參考
-# False 會以實際總成長% 計算，提供若整個時段都持有不賣的最佳配置 
+            # Append to the results list
+            results.append(pd.DataFrame([x for x in result['ResultSet']['Result']]))
 
-fig, pred_port, growth = EfficientFrontier(table, annualized=True) 
-display(pred_port)
-print('\n')
-print(f'{start} - {end} 總成長(%)')
-display(growth)
-print('\n')
-display(fig)
+        # Return a concatenated Pandas Dataframe of results list
+        return pd.concat(results)
 
-# 前提一、回報與波動為假設買後不動
-# 前提二、過去表現不代表未來表現
-# 前提三、此測試不考慮手續費、稅務等投資時重要考量
+
+
+#####################
+# Under Development #
+#####################
+# def getport(start, end, symbols):
+#   '''
+#   A function that collects all data from a list of assets from Yahoo! Finance
+#   '''
+#   assets = [get_symbol(x) for x in symbols]
+#   valid_date(start)
+#   valid_date(end)
+#   assert isinstance(symbols, list), '代號必須是list of string'
+#   for symbol in symbols:
+#       assert isinstance(symbol, str), '代號必須是list of string'
+#   assert isinstance(assets, list), '資產名稱必須是list of string'
+#   for asset in assets:
+#       assert isinstance(asset, str), '資產名稱必須是list of string'
+#   assert len(symbols) == len(assets), 'symbols 和assets 必須是同數量'
+#   # Define date_range
+#   date_range = pd.date_range(
+#     start = start,
+#     end = end,  
+#     freq = 'B', 
+#     normalize = True
+#     )
+#   df_list = [yf.Ticker(symbol).history(start=start, end=end) for symbol in symbols]
+#   for df, asset in zip(df_list, assets):
+#     col_list = df.drop('Close', axis = 1).columns
+#     df.drop(col_list, axis=1, inplace=True)
+#     df.rename(columns={'Close': asset}, inplace=True)
+#   df = pd.concat(df_list, axis=1, ignore_index=False)
+#   df = df.fillna(method='ffill').reindex(index=date_range, method='ffill')
+#   df.dropna(how='any', inplace=True)
+#   return df
+
+# def EfficientFrontier(table, annualized):
+#     ### Populate portfolios
+# # Calculate daily and annual returns of the stocks
+#     assert isinstance(table, pd.DataFrame), "table 必須是DataFrame 格式"
+#     assert isinstance(annualized, bool), "annualized 必須是True 或是False"
+#     if annualized == False:
+#       returns_daily = table.pct_change()
+#       returns_total = (table.iloc[-1]-table.iloc[0])/table.iloc[0]
+
+#       # Get daily and covariance of returns of the stock
+#       cov_daily = returns_daily.cov()
+#       cov_total = cov_daily * table.shape[0]
+      
+#     elif annualized == True:
+#       returns_daily = table.pct_change()
+#       returns_total = returns_daily.mean() * 252
+
+#       # Get daily and covariance of returns of the stock
+#       cov_daily = returns_daily.cov()
+#       cov_total = cov_daily * 252
+#     else:
+#       raise ValueError("annualized 必須是True 或是False")
+
+#     # Define selected assets
+#     selected = table.columns
+
+#     # Empty lists to store returns, volatility and weights of imiginary portfolios
+#     port_returns = []
+#     port_volatility = []
+#     sharpe_ratio = []
+#     stock_weights = []
+
+#     # Set the number of combinations for imaginary portfolios
+#     num_assets = len(selected)
+#     num_portfolios = 50000
+
+#     # Set random seed for reproduction's sake
+#     np.random.seed(101)
+
+#     # Populate the empty lists with each portfolios returns,risk and weights
+#     for single_portfolio in range(num_portfolios):
+#         weights = np.random.random(num_assets)
+#         weights /= np.sum(weights)
+#         returns = np.dot(weights, returns_total)
+#         volatility = np.sqrt(np.dot(weights.T, np.dot(cov_total, weights)))
+#         sharpe = returns / volatility
+#         sharpe_ratio.append(sharpe)
+#         port_returns.append(returns)
+#         port_volatility.append(volatility)
+#         stock_weights.append(weights)
+
+#     # A dictionary for Returns and Risk values of each portfolio
+#     portfolio = {'回報': port_returns,
+#                 '波動': port_volatility,
+#                 '夏普比例': sharpe_ratio}
+
+#     # Extend original dictionary to accomodate each ticker and weight in the portfolio
+#     for counter,symbol in enumerate(selected):
+#         portfolio[symbol+' 權重'] = [Weight[counter] for Weight in stock_weights]
+
+#     # Make a nice dataframe of the extended dictionary
+#     df = pd.DataFrame(portfolio)
+
+#     # Get better labels for desired arrangement of columns
+#     column_order = ['回報', '波動', '夏普比例'] + [stock+' 權重' for stock in selected]
+
+#     # Reorder dataframe columns
+#     df = df[column_order]
+#     df.dropna(how="any", inplace=True, axis=0)
+#     min_volatility = df['波動'].min()
+#     max_sharpe = df['夏普比例'].max()
+
+#     # use the min, max values to locate and create the two special portfolios
+#     sharpe_portfolio = df.loc[df['夏普比例'] == max_sharpe]
+#     min_variance_port = df.loc[df['波動'] == min_volatility]
+
+#     mv_port = min_variance_port.copy()
+#     ms_port = sharpe_portfolio.copy()
+#     for col in min_variance_port.columns:
+#         if col == '夏普比例':
+#             pass
+#         else:
+#             mv_port[col] = mv_port[col].apply(lambda x: x*100)
+            
+#     for col in sharpe_portfolio.columns:
+#         if col == '夏普比例':
+#             pass
+#         else:
+#             ms_port[col] = ms_port[col].apply(lambda x: x*100)
+#     # plot frontier, max sharpe & min Volatility values with a scatterplot
+#     fig = px.scatter(
+#         df, 
+#         x="波動", 
+#         y="回報",  
+#         size_max=60,
+#         color="夏普比例",
+#         color_continuous_scale = 'aggrnyl',
+#         hover_data=df.columns
+        
+#     )
+#     fig.add_scatter(
+#         x=sharpe_portfolio['波動'], 
+#         y=sharpe_portfolio['回報'],
+#         mode="markers",
+#         marker={'size': 20,'color': "red",},
+#         showlegend=False
+#     )
+#     fig.add_scatter(
+#         x=min_variance_port['波動'], 
+#         y=min_variance_port['回報'],
+#         mode="markers",
+#         marker={'size': 20,'color': "blue",},
+#         showlegend=False
+#     )
+#     fig.update_traces(textposition='top center')
+
+#     fig.update_layout(
+#         height=600,
+#         width=1000,
+#         title_text='Portfolio'
+#     )
+#     pred_port = pd.concat([mv_port.T, ms_port.T], axis = 1)
+#     pred_port.columns = ['最低波動', '最高夏普']
+#     growth = (((table.iloc[-1]-table.iloc[0])/table.iloc[0])*100)
+#     return fig, pred_port, growth
+
+# # Codes modified from: https://medium.com/python-data/efficient-frontier-portfolio-optimization-with-python-part-2-2-2fe23413ad94
+
+# # 定義開始日期
+# start = '2010-01-01'
+
+# # 定義結束日期
+# end = '2019-12-31'
+
+# # 投資組合裡面的所有資產，必須是Yahoo! Finance 英文版上的代號，必須是在英文的引號("")裡面
+# # 請上https://finance.yahoo.com/ 查詢代號
+# symbols = ["GOOG",
+#            "AAPL",
+#            "FB",
+#            "AMZN",
+#            "TWD=X",
+#            "^IRX",
+#            "^FVX",
+#            "^TNX",
+#            "^TYX",
+#            "WTI",
+#            "USO",
+#            "GLDM"
+#            ]
+
+# table = getport(start, end, symbols)
+
+# # annualized 是年化與否
+# # True 會以年化平均計算，提供未來配置的參考
+# # False 會以實際總成長% 計算，提供若整個時段都持有不賣的最佳配置 
+
+# fig, pred_port, growth = EfficientFrontier(table, annualized=True) 
+# display(pred_port)
+# print('\n')
+# print(f'{start} - {end} 總成長(%)')
+# display(growth)
+# print('\n')
+# display(fig)
+
+# # 前提一、回報與波動為假設買後不動
+# # 前提二、過去表現不代表未來表現
+# # 前提三、此測試不考慮手續費、稅務等投資時重要考量
+
+############
+# Rulesets #
+############
+# def get_leverage(self):
+#     '''
+#     A custom
+#     '''
+#     self.leverage = set_leverage(self.balance)
+#     print(f"Current Leverage: {self.leverage}")
+
+# def get_investable(self):
+#     return self.balance * self.leverage