+++ CHANGELOG.md
@@ -8,8 +8,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - `data.py` to acquire data through API
 - `feature.py` to perform feature engineering and include desired datasets
+- `portfolio.py`:
+    - Trade: Buy
+    - Trade: Sell
+    - Trade: Fee
 
-## [0.0.1] - 2014-05-31
+## [0.1.0] - 2021-03-27
+### Added
+- `Portfolio.search()` that queries keywords on Yahoo! Finance autocomplete API
+- Added Docstrings to `portfolio.py`
+
+### Changed
+- In `Portfolio()` class, `self.balance` changed to `self.cash` to allow trade features
+
+## [0.0.1] - 2021-03-26
 ### Added
 - This CHANGELOG file to keep track of changes happening around this long-term project
 - README to indicate that this project is based on a previous project